+++ CHANGELOG.md
@@ -5,13 +5,13 @@
 - coverage - add macros `/*coverage-disable*/` and `/*coverage-enable*/`.
 - jslint - add html and css linting back into jslint.
 - jslint - add new warning requiring paren around plus-separated concatenations.
-- jslint - relax warning against console.log and friends and deprecate directive `devel`
 - jslint - require regexp to use open-form.
 - jslint - try to improve parser to be able to parse jquery.js without stopping.
 - jslint - unify analysis of variable-assignment/function-parameters into one function
-- jslint - add new warning "Expected Object.create(null) instead of {}"
 
 # v2023.1.1-beta
+- jslint-ci - revamp auto-updating and add shell-function shGithubCheckoutRemote
+- test - print time-finished after test-run
 - jslint - hide warning about unordered case-statements behind beta-flag
 - ci - bugfix - update shell-function shCiBase() to handle undefined fileMain
 - ci - auto-update version-number in main mjs-module